fix(tests): make googletest a submodule.
This commit makes googletest a submodule rather than declare as
a dependency using cmake.
It also introduces all the necessary cmake changes to make this
happen.

All the docs files will be removed from the repository.
Instead, their building will be incorporated into cmake or otherwise
build instructions for them will be included in the README.
It is really absurd to have the documentation built as part of the
repository.