cmake(Presets): convert from `CMakeSettings.json` to `CMakePresets.json`.
This commit also eliminates use of readthedocs from the main branch.
This is due to difficulities running cmake and generating the project
on the linux server of readthedocs.

The solution I will use is to create a branch for each documentation
version, or to create a branch for documentation in general.
If I use the latter approach, I will probably have to delete that branch
each time a new version emerges. I will try the former approach first.