fix(examples): ensure examples compile and run.
Why:
The API of the library has changed, however, the examples were not
adapted to the new API. In particular, the names of data structures
used by the examples weren't changed. This cause failure of compilation.

Also, there were some inconveniences with `run_doxygen.cmake`.

How:
Change the names of the data structures to reflect the new API.

Resolve the incovneniences in `run_doxygen.cmake`.