feat!: create new `set` function.
Why?
The purpose of this function is to facilitate filling a state machine
table, by sequentially filling in the state of a simple regular express-
ion.
Tests have also been created for this function.

BREAKING_CHANGE Acess to the start state and dead states of the machine
is now thorugh two functions, `getStartState` and `getDeadState`.
These two are static functions of the `FiniteStateMachine` class (prev-
iously, you could access the dead state from a specialized class, but
now you cannot.)